updated precision

+++ functime/conversion.py
@@ -12,7 +12,7 @@ def df_to_ndarray(df: pl.DataFrame) -> np.ndarray:
 def X_to_numpy(X: pl.DataFrame) -> np.ndarray:
     X_arr = (
         X.lazy()
-        .select(pl.col(X.columns[2:]).cast(pl.Float32))
+        .select(pl.col(X.columns[2:]).cast(pl.Float64))
         .select(
             pl.when(pl.all().is_infinite() | pl.all().is_nan())
             .then(None)
@@ -30,7 +30,7 @@ def X_to_numpy(X: pl.DataFrame) -> np.ndarray:
 def y_to_numpy(y: pl.DataFrame) -> np.ndarray:
     y_arr = (
         y.lazy()
-        .select(pl.col(y.columns[-1]).cast(pl.Float32))
+        .select(pl.col(y.columns[-1]).cast(pl.Float64))
         .select(
             pl.when(pl.all().is_infinite() | pl.all().is_nan())
             .then(None)
